refactor: UrlSearchBar.swift
